Drop logrus and modernize codebase
Replace logrus with log/slog, removing the dependency. Split modiff.go
into focused files, use slices.SortFunc, fix type alias, remove dead
CheckURLValid export, switch to built-in shell completion, and add
cmd/ test coverage. Also fix generateCompareLink to verify both
modules share a known host, cap subsection header level to valid
Markdown, and handle GOPROXY pipe separator.
