+++ .github/workflows/ci.yml
@@ -55,15 +55,6 @@ jobs:
           files: build/coverage/coverprofile
           token: ${{ secrets.CODECOV_TOKEN }}
 
-  verify-completions:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-      - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7.0.0
-        with:
-          go-version-file: go.mod
-      - run: make verify-completions
-
   verify-tidy:
     runs-on: ubuntu-latest
     steps:
