+++ go.mod
@@ -6,7 +6,6 @@ require (
 	github.com/onsi/ginkgo/v2 v2.32.1
 	github.com/onsi/gomega v1.42.0
 	github.com/saschagrunert/ccli/v3 v3.0.0
-	github.com/sirupsen/logrus v1.9.4
 	github.com/urfave/cli/v3 v3.10.0
 )
